# Conflicts:
#	app/src/main/java/com/xjek/provider/views/dashboard/DashBoardActivity.kt
#	app/src/main/java/com/xjek/provider/views/dashboard/DashBoardViewModel.kt
#	base/src/main/java/com/xjek/base/base/BaseActivity.kt
#	base/src/main/java/com/xjek/base/data/Constants.kt
#	base/src/main/java/com/xjek/base/location_service/BaseLocationService.java
#	taxiservice/src/main/java/com/xjek/taxiservice/views/main/TaxiDashboardActivity.kt

+++ app/src/main/java/com/xjek/provider/models/ProfileData.kt
@@ -35,7 +35,7 @@ class ProfileData {
     @SerializedName("wallet_balance")
     var walletBalance: Int? = null
     @SerializedName("rating")
-    var rating: Int? = null
+    var rating: Float? = null
     @SerializedName("status")
     var status: String? = null
     @SerializedName("admin_id")
